+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# :package: 0.3.0
+
+01. :gift: Added Apache `beanutils` converters
+    01. :gift: Added class `dev.orne.beans.converters.UriConverter`
+    01. :gift: Added methods `registerNetConversors` to `dev.orne.beans.converters.OrneBeansConverters`
+
 # :package: 0.2.0
 
 01. Improved class `AbstractIdentity`